[Update] APIHandlerTests moved to handlers package in test file
