+++ src/test/java/com/cebix/investmenttrackerapp/handlers/MarketStatusAPIHandlerTests.java
@@ -1,6 +1,5 @@
-package com.cebix.investmenttrackerapp;
+package com.cebix.investmenttrackerapp.handlers;
 
-import com.cebix.investmenttrackerapp.handlers.MarketStatusAPIHandler;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
