+++ src/test/java/com/cebix/investmenttrackerapp/handlers/StocksAPIHandlerTests.java
@@ -1,12 +1,10 @@
-package com.cebix.investmenttrackerapp;
+package com.cebix.investmenttrackerapp.handlers;
 
-import com.cebix.investmenttrackerapp.handlers.StocksAPIHandler;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import reactor.core.publisher.Mono;
 import reactor.test.StepVerifier;
 
 import java.io.IOException;
